Refactor recordLoginActivity into ServerConnectivityRepository
- Moved HTTP call logic for recordLoginActivity from MyPlanetLite to ServerConnectivityRepository.
- Inject CoroutineDispatcher to the repository and use it for the suspend method.
- Re-use the OkHttpClient initialization without caching it as a property if not used elsewhere, or directly instantiate the builder if it's purely for DI passing.
- Added tests for the new method in ServerConnectivityRepositoryTest using MockWebServer to assert POST body and headers.
